requirements.txt
moderngl>=5.7
glfw>=2.6.2
numpy>=1.26
av>=11.0
ffmpeg-python>=0.2.0
Pillow>=10.0
playwright>=1.47.0

